# packages/settlrl-app/frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Settlrl Render",
  "version": "0.1.0"
 },
 "paths": {
  "/api/preview": {
   "get": {
    "summary": "Get Preview",
    "description": "The board a new game would open on, for the map picker \u2014 no game is\ncreated. Terrain and ports depend only on the seed (not the placement).",
    "operationId": "get_preview_api_preview_get",
    "parameters": [
     {
      "name": "seed",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 0,
       "title": "Seed"
      }
     },
     {
      "name": "n_players",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 4,
       "title": "N Players"
      }
     },
     {
      "name": "number_placement",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "random",
        "spiral"
       ],
       "type": "string",
       "default": "random",
       "title": "Number Placement"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BoardModel"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/games": {
   "post": {
    "summary": "Post Create",
    "description": "Create a game, or return the caller's place in line when the server\nis at its concurrency cap (a ``202`` they re-POST with ``ticket``).",
    "operationId": "post_create_api_games_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_CreateRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "anyOf": [
          {
           "$ref": "#/components/schemas/_CreatedModel"
          },
          {
           "$ref": "#/components/schemas/_QueuedModel"
          }
         ],
         "title": "Response Post Create Api Games Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/api/games/{game_id}/join": {
   "post": {
    "summary": "Post Join",
    "operationId": "post_join_api_games__game_id__join_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "game_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Game Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_JoinRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/_JoinedModel"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/games/{game_id}": {
   "get": {
    "summary": "Get Game",
    "operationId": "get_game_api_games__game_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "game_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Game Id"
      }
     },
     {
      "name": "X-Seat-Tokens",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Seat-Tokens"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GameModel"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/games/{game_id}/action": {
   "post": {
    "summary": "Post Action",
    "description": "Apply the acting seat's move; the request must prove that seat.",
    "operationId": "post_action_api_games__game_id__action_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "game_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Game Id"
      }
     },
     {
      "name": "X-Seat-Tokens",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Seat-Tokens"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_ActionRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GameModel"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/games/{game_id}/events": {
   "get": {
    "summary": "Get Events",
    "description": "Server-sent events: the requester's snapshot now, then again on\nevery state change (moves, bot plays, chat, joins). Comment lines\nkeep idle connections alive.",
    "operationId": "get_events_api_games__game_id__events_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "game_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Game Id"
      }
     },
     {
      "name": "X-Seat-Tokens",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Seat-Tokens"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/games/{game_id}/chat": {
   "post": {
    "summary": "Post Chat",
    "description": "Append a chat message to the game log.",
    "operationId": "post_chat_api_games__game_id__chat_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "game_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Game Id"
      }
     },
     {
      "name": "X-Seat-Tokens",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Seat-Tokens"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_ChatRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GameModel"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/games/{game_id}/record": {
   "get": {
    "summary": "Get Record",
    "description": "The finished game as ``GameRecord`` JSON -- a self-contained,\nreplayable transcript. 409 while running: replaying a record\nreconstructs hidden hands, so live games don't export.",
    "operationId": "get_record_api_games__game_id__record_get",
    "parameters": [
     {
      "name": "game_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Game Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/games/{game_id}/replay": {
   "post": {
    "summary": "Post Replay From Game",
    "description": "Load a finished game for replay (409 while it is still running:\nreplaying reconstructs hidden hands).",
    "operationId": "post_replay_from_game_api_games__game_id__replay_post",
    "parameters": [
     {
      "name": "game_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Game Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReplayStateModel"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "BankModel": {
    "properties": {
     "resources": {
      "$ref": "#/components/schemas/ResourceCounts"
     },
     "dev_cards": {
      "type": "integer",
      "title": "Dev Cards"
     }
    },
    "type": "object",
    "required": [
     "resources",
     "dev_cards"
    ],
    "title": "BankModel",
    "description": "Cards left in the supply: resource stacks plus the development deck."
   },
...
   "BoardModel": {
    "properties": {
     "tiles": {
      "items": {
       "$ref": "#/components/schemas/TileModel"
      },
      "type": "array",
      "title": "Tiles"
     },
     "buildings": {
      "items": {
       "$ref": "#/components/schemas/BuildingModel"
      },
      "type": "array",
      "title": "Buildings",
      "default": []
     },
     "roads": {
      "items": {
       "$ref": "#/components/schemas/RoadModel"
      },
      "type": "array",
      "title": "Roads",
      "default": []
     },
     "ports": {
      "items": {
       "$ref": "#/components/schemas/PortModel"
      },
      "type": "array",
      "title": "Ports",
      "default": []
     },
     "players": {
      "items": {
       "$ref": "#/components/schemas/PlayerModel"
      },
      "type": "array",
      "title": "Players",
      "default": []
     },
     "robber": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/HexModel"
       },
       {
        "type": "null"
       }
      ]
     },
     "bank": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/BankModel"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "tiles"
    ],
    "title": "BoardModel"
   },
...
   "BuildingModel": {
    "properties": {
     "cube": {
      "$ref": "#/components/schemas/CubeModel"
     },
     "player": {
      "type": "integer",
      "title": "Player"
     },
     "kind": {
      "type": "string",
      "enum": [
       "settlement",
       "city"
      ],
      "title": "Kind"
     }
    },
    "type": "object",
    "required": [
     "cube",
     "player",
     "kind"
    ],
    "title": "BuildingModel",
    "description": "A settlement or city sitting on a vertex, owned by a 0-indexed player."
   },
...
   "GameModel": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "version": {
      "type": "integer",
      "title": "Version",
      "default": 0
     },
     "board": {
      "$ref": "#/components/schemas/BoardModel"
     },
     "status": {
      "$ref": "#/components/schemas/GameStatusModel"
     },
     "actions": {
      "items": {
       "$ref": "#/components/schemas/ActionModel"
      },
      "type": "array",
      "title": "Actions",
      "default": []
     },
     "bot_move": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/BotMoveModel"
       },
       {
        "type": "null"
       }
      ]
     },
     "log": {
      "items": {
       "$ref": "#/components/schemas/LogEntryModel"
      },
      "type": "array",
      "title": "Log",
      "default": []
     },
     "belief": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/BeliefModel"
       },
       {
        "type": "null"
       }
      ]
     },
     "seats_claimed": {
      "items": {
       "type": "integer"
      },
      "type": "array",
      "title": "Seats Claimed",
      "default": []
     },
     "your_seats": {
      "items": {
       "type": "integer"
      },
      "type": "array",
      "title": "Your Seats",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "id",
     "board",
     "status"
    ],
    "title": "GameModel",
    "description": "One requester's snapshot: board + status + their legal moves.\n\nViews are per-seat: ``your_turn`` / ``actions`` / unredacted hands /\n``belief`` follow the seats the request's tokens prove."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "HexModel": {
    "properties": {
     "q": {
      "type": "integer",
      "title": "Q"
     },
     "r": {
      "type": "integer",
      "title": "R"
     }
    },
    "type": "object",
    "required": [
     "q",
     "r"
    ],
    "title": "HexModel",
    "description": "An axial (pointy-top) tile coordinate."
   },
...
   "PlayerModel": {
    "properties": {
     "player": {
      "type": "integer",
      "title": "Player"
     },
     "resource_cards": {
      "type": "integer",
      "title": "Resource Cards"
     },
     "dev_cards": {
      "type": "integer",
      "title": "Dev Cards"
     },
     "victory_points": {
      "type": "integer",
      "title": "Victory Points"
     },
     "knights_played": {
      "type": "integer",
      "title": "Knights Played",
      "default": 0
     },
     "longest_road": {
      "type": "boolean",
      "title": "Longest Road",
      "default": false
     },
     "largest_army": {
      "type": "boolean",
      "title": "Largest Army",
      "default": false
     },
     "resources": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ResourceCounts"
       },
       {
        "type": "null"
       }
      ]
     },
     "dev_card_types": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/DevCardCounts"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "player",
     "resource_cards",
     "dev_cards",
     "victory_points"
    ],
    "title": "PlayerModel",
    "description": "Summary stats for one (0-indexed) player.\n\nThe per-type breakdowns are private: the server nulls them out for every\nseat the requester doesn't own (counts stay public, as at a real table)."
   },
   "PortModel": {
    "properties": {
     "a": {
      "$ref": "#/components/schemas/CubeModel"
     },
     "b": {
      "$ref": "#/components/schemas/CubeModel"
     },
     "resource": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "sheep",
         "wheat",
         "wood",
         "brick",
         "ore"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Resource"
     }
    },
    "type": "object",
    "required": [
     "a",
     "b",
     "resource"
    ],
    "title": "PortModel",
    "description": "A harbour spanning two coastal vertices.\n\n``resource`` is the resource of a 2:1 port, or ``None`` for a 3:1 general\nport."
   },
   "ReplayStateModel": {
    "properties": {
     "move": {
      "type": "integer",
      "title": "Move"
     },
     "n_moves": {
      "type": "integer",
      "title": "N Moves"
     },
     "board": {
      "$ref": "#/components/schemas/BoardModel"
     },
     "log": {
      "items": {
       "$ref": "#/components/schemas/LogEntryModel"
      },
      "type": "array",
      "title": "Log",
      "default": []
     },
     "winner": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Winner"
     },
     "seats": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Seats"
     }
    },
    "type": "object",
    "required": [
     "move",
     "n_moves",
     "board"
    ],
    "title": "ReplayStateModel",
    "description": "The Replay view's snapshot after ``move`` of ``n_moves`` moves.\n\n``log`` holds the moves played up to that point (the win line appears only\nat the final move); ``winner`` / ``seats`` describe the whole record."
   },
...
   "RoadModel": {
    "properties": {
     "a": {
      "$ref": "#/components/schemas/CubeModel"
     },
     "b": {
      "$ref": "#/components/schemas/CubeModel"
     },
     "player": {
      "type": "integer",
      "title": "Player"
     }
    },
    "type": "object",
    "required": [
     "a",
     "b",
     "player"
    ],
    "title": "RoadModel",
    "description": "A road along an edge between two vertices, owned by a 0-indexed player."
   },
   "Terrain": {
    "type": "string",
    "enum": [
     "wheat",
     "sheep",
     "wood",
     "ore",
     "brick",
     "desert"
    ],
    "title": "Terrain"
   },
   "TileModel": {
    "properties": {
     "q": {
      "type": "integer",
      "title": "Q"
     },
     "r": {
      "type": "integer",
      "title": "R"
     },
     "terrain": {
      "$ref": "#/components/schemas/Terrain"
     },
     "number": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Number"
     }
    },
    "type": "object",
    "required": [
     "q",
     "r",
     "terrain"
    ],
    "title": "TileModel"
   },
...
   "_ActionRequest": {
    "properties": {
     "flat": {
      "type": "integer",
      "title": "Flat"
     }
    },
    "type": "object",
    "required": [
     "flat"
    ],
    "title": "_ActionRequest"
   },
   "_ChatRequest": {
    "properties": {
     "text": {
      "type": "string",
      "title": "Text"
     },
     "player": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Player"
     }
    },
    "type": "object",
    "required": [
     "text"
    ],
    "title": "_ChatRequest"
   },
   "_CreateRequest": {
    "properties": {
     "seed": {
      "type": "integer",
      "title": "Seed",
      "default": 0
     },
     "n_players": {
      "type": "integer",
      "enum": [
       2,
       4
      ],
      "title": "N Players",
      "default": 4
     },
     "number_placement": {
      "type": "string",
      "enum": [
       "random",
       "spiral"
      ],
      "title": "Number Placement",
      "default": "random"
     },
     "seats": {
      "anyOf": [
       {
        "items": {
         "anyOf": [
          {
           "type": "string"
          },
          {
           "$ref": "#/components/schemas/_SeatSpec"
          }
         ]
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Seats"
     },
     "claim": {
      "type": "string",
      "enum": [
       "all",
       "first",
       "none"
      ],
      "title": "Claim",
      "default": "all"
     },
     "ticket": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ticket"
     }
    },
    "type": "object",
    "title": "_CreateRequest"
   },
   "_CreatedModel": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "seats": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Seats"
     },
     "tokens": {
      "additionalProperties": {
       "type": "string"
      },
      "type": "object",
      "title": "Tokens"
     }
    },
    "type": "object",
    "required": [
     "id",
     "seats",
     "tokens"
    ],
    "title": "_CreatedModel",
    "description": "A fresh game: its id and the creator's seat tokens."
   },
   "_JoinRequest": {
    "properties": {
     "seat": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Seat"
     }
    },
    "type": "object",
    "title": "_JoinRequest"
   },
   "_JoinedModel": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "seat": {
      "type": "integer",
      "title": "Seat"
     },
     "token": {
      "type": "string",
      "title": "Token"
     }
    },
    "type": "object",
    "required": [
     "id",
     "seat",
     "token"
    ],
    "title": "_JoinedModel"
   },
...
   "_QueuedModel": {
    "properties": {
     "queued": {
      "type": "boolean",
      "const": true,
      "title": "Queued",
      "default": true
     },
     "ticket": {
      "type": "string",
      "title": "Ticket"
     },
     "position": {
      "type": "integer",
      "title": "Position"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     }
    },
    "type": "object",
    "required": [
     "ticket",
     "position",
     "total"
    ],
    "title": "_QueuedModel",
    "description": "The server is at its concurrency cap: the caller's place in line. They\nre-POST with ``ticket`` until they get a :class:`_CreatedModel` back."
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "api/auth/login"
     }
    }
   }
  }
 }
}